refactor logins to go from user to login component
