+++ src/Components/User/User.js
@@ -6,58 +6,36 @@ import * as actions from '../../Actions/';
 import './User.css';
 import clearLocalStorage from '../../helpers/clearLocalStorage';
 import { NavLink } from 'react-router-dom';
+import getKeyFromLS from '../../helpers/getKeyFromLS';
 
 class User extends Component {
   constructor() {
     super();
   }
-  componentDidMount() {
-    this.checkForKey();    
+
+  componentDidMount = async () => {
+  	await this.loadUser();
   }
 
-  async sendKey (key) {
-    const loginResponse = await fetch('http://localhost:3000/api/v1/login', {
-      method: 'GET',
-      headers: {
-        "x-token": key,
-        "CONTENT-TYPE": 'application/json'
-      }
-    });
-    const userData = await loginResponse.json();
+  loadUser = async () => {
+  	try {
+	    const user = await fetch('http://localhost:3000/api/v1/users', {
+	      method: 'GET',
+	      headers: {
+	        "x-token": getKeyFromLS(),
+	        "CONTENT-TYPE": 'application/json'
+	      }
+	    });
+	    const userData = await user.json();
 
-    this.props.updateUser(userData[0]);
-    this.props.history.push('/user');
-  }
+	    this.props.updateUser(userData[0]);
 
-  checkForKey() {
-    const userKey = JSON.parse(localStorage.getItem('123rtx-token'));
-    if (this.props.location.search.length > 500) {
-      const passedToken = this.getToken();
-      localStorage.setItem('123rtx-token', JSON.stringify(passedToken));
-      this.sendKey(passedToken);
-    } else if (userKey) {
-      this.sendKey(userKey);
-    } else {
-    	clearLocalStorage();
+  	} catch (e) {
+  		console.log('heeeeeeeyo')
     	window.location="https://tr-personal-proj.e1.loginrocket.com"
-    }
-
-    //pass to backend
-    // fetch('/api/v1/login', {
-    //   headers: {
-    //     x-token: 
-    //   }
-    // })
+    	return
+  	}
   }
-
-  getToken () {
-    const tokenRegex = new RegExp(/\?token=/);
-    const parseToken = this.props.location.search.replace(tokenRegex, '');
-
-    return parseToken;
-  }
-
-
 
   render() {
     return (
