+++ src/helpers/getKeyFromLS.js
@@ -1,0 +1,6 @@
+const getKeyFromLS = () => {
+	const userKey = JSON.parse(localStorage.getItem('123rtx-token'));
+	return userKey
+}
+
+export default getKeyFromLS;
